Split the script into two plus helpers
